Change a size_t to a uint64_t in the forward declaration of _hpcrun_backtrace() in unwind/common/backtrace.c to match the function definition.  This was breaking the build on BlueGene since size_t != uint64_t on 32-bit.

+++ src/tool/hpcrun/unwind/common/backtrace.c
@@ -33,7 +33,7 @@
 
 static csprof_cct_node_t*
 _hpcrun_backtrace(csprof_state_t* state, ucontext_t* context,
-		  int metricId, size_t metricIncr,
+		  int metricId, uint64_t metricIncr,
 		  int skipInner);
